moved all test files into test folders.
