+++ src/Tests/App.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import App from './components/App';
+import App from '../components/App';
 
 test('this test is empty until we sort some stuff out', () => {
   // render(<App />);
