+++ src/Tests/ProductDetails.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import ProductDetail from './components/ProductDetail';
+import ProductDetail from '../components/ProductDetail';
 
 test('is able to render the ProductDetails component', () => {
   render(<ProductDetail />);
